Move wallet connection delegate component
