+++ apps/voting/src/shared/wallet-connection.tsx
@@ -3,10 +3,21 @@ import {
     CONCORDIUM_WALLET_CONNECT_PROJECT_ID,
     WalletConnectConnector,
     WalletConnection,
+    WalletConnectionDelegate,
     WalletConnector,
 } from '@concordium/wallet-connectors';
 import { SignClientTypes } from '@walletconnect/types';
-import { PropsWithChildren, createContext, useCallback, useContext, useEffect, useMemo, useState } from 'react';
+import {
+    Component,
+    PropsWithChildren,
+    createContext,
+    useCallback,
+    useContext,
+    useEffect,
+    useMemo,
+    useState,
+} from 'react';
+import { NETWORK } from './constants';
 
 export const WALLET_CONNECT_OPTS: SignClientTypes.Options = {
     projectId: CONCORDIUM_WALLET_CONNECT_PROJECT_ID,
@@ -128,7 +139,7 @@ function WalletProvider({ connector, children }: WalletProviderProps) {
     return <Provider value={contextValue}>{children}</Provider>;
 }
 
-export type WalletsProviderProps = PropsWithChildren<{
+type WalletsProviderProps = PropsWithChildren<{
     /** Connector instance to the Concordium browser wallet */
     browser: BrowserWalletConnector;
     /** Connector instance to wallet connect compatible Concordium wallet */
@@ -141,7 +152,7 @@ export type WalletsProviderProps = PropsWithChildren<{
  * Component whose sole purpose is to provide connection management functionality to the component tree below the
  * component acting as wallet connector delegate.
  */
-export function WalletsProvider({ browser, walletConnect, activeWallet, children }: WalletsProviderProps) {
+function WalletsProvider({ browser, walletConnect, activeWallet, children }: WalletsProviderProps) {
     return (
         <activeWalletContext.Provider value={activeWallet}>
             <WalletProvider connector={browser}>
@@ -149,4 +160,114 @@ export function WalletsProvider({ browser, walletConnect, activeWallet, children
             </WalletProvider>
         </activeWalletContext.Provider>
     );
+}
+
+function updateMapEntry<K, V>(map: Map<K, V>, key: K | undefined, value: V | undefined) {
+    const res = new Map(map);
+    if (key !== undefined) {
+        if (value !== undefined) {
+            res.set(key, value);
+        } else {
+            res.delete(key);
+        }
+    }
+    return res;
+}
+
+type WalletConnectionManagerProps = PropsWithChildren;
+interface WalletConnectionManagerState {
+    browserWalletConnector: BrowserWalletConnector | undefined;
+    walletConnectConnector: WalletConnectConnector | undefined;
+    connections: WalletConnection[];
+    accounts: Map<WalletConnection, string | undefined>;
+    chains: Map<WalletConnection, string | undefined>;
+}
+
+/**
+ * Manages connections for the different wallets the application integrates with. Connection details will be available to
+ * the component tree below this component.
+ */
+export class WalletConnectionManager
+    extends Component<WalletConnectionManagerProps, WalletConnectionManagerState>
+    implements WalletConnectionDelegate {
+    constructor(props: WalletConnectionManagerProps) {
+        super(props);
+
+        this.state = {
+            browserWalletConnector: undefined,
+            walletConnectConnector: undefined,
+            connections: [],
+            accounts: new Map(),
+            chains: new Map(),
+        };
+    }
+    onChainChanged(connection: WalletConnection, genesisHash: string): void {
+        this.setState((state) => ({
+            ...state,
+            chains: updateMapEntry(state.chains, connection, genesisHash),
+        }));
+    }
+    onAccountChanged(connection: WalletConnection, address: string | undefined): void {
+        this.setState((state) => ({
+            ...state,
+            accounts: updateMapEntry(state.accounts, connection, address),
+        }));
+    }
+    onConnected(connection: WalletConnection, address: string | undefined): void {
+        this.setState((state) => {
+            // The first entry is the active connection.
+            const connections = [connection, ...state.connections.filter((c) => c === connection)];
+            return {
+                ...state,
+                connections,
+                accounts: updateMapEntry(state.accounts, connection, address),
+            };
+        });
+    }
+    onDisconnected(connection: WalletConnection): void {
+        this.setState((state) => {
+            const connections = state.connections.filter((c) => c !== connection);
+            return {
+                ...state,
+                connections,
+                accounts: updateMapEntry(state.accounts, connection, undefined),
+                chains: updateMapEntry(state.chains, connection, undefined),
+            };
+        });
+    }
+
+    async componentDidMount(): Promise<void> {
+        const bwPromise = BrowserWalletConnector.create(this);
+        const wcPromise = WalletConnectConnector.create(WALLET_CONNECT_OPTS, this, NETWORK);
+        const [bw, wc] = await Promise.all([bwPromise, wcPromise]);
+
+        this.setState({ browserWalletConnector: bw, walletConnectConnector: wc });
+    }
+
+    render() {
+        const { browserWalletConnector, walletConnectConnector } = this.state;
+        if (browserWalletConnector === undefined || walletConnectConnector === undefined) {
+            return null;
+        }
+
+        const connection = this.state.connections[0];
+        const activeWallet: ActiveWallet =
+            connection === undefined
+                ? {}
+                : {
+                    chain: this.state.chains.get(connection),
+                    account: this.state.accounts.get(connection),
+                    connection,
+                };
+
+        return (
+            <WalletsProvider
+                browser={browserWalletConnector}
+                walletConnect={walletConnectConnector}
+                activeWallet={activeWallet}
+            >
+                {this.props.children}
+            </WalletsProvider>
+        );
+    }
 }
